WIP: docs/briefs/t-substrate-nominal-opaque-secret-worker.md

+++ src/v3/compiler/src/lens_testgen.rs
@@ -424,7 +424,7 @@ impl<'a> TestgenLens<'a> {
             vec![
                 FieldValue::Literal(LiteralBits::String(bind_name.to_string())),
                 self.sum_variant("ComparisonOp", comparator, Vec::new()),
-                FieldValue::Literal(LiteralBits::Int(bound as i64)),
+                FieldValue::Literal(LiteralBits::Int(i128::from(bound as i64))),
             ],
         )
     }
